♻️ Move tx.go to internal/wrap and add alias to it
